+++ kube/examples/crd_reflector.rs
@@ -6,8 +6,7 @@ use std::time::Duration;
 
 use kube::{
     api::{ListParams, Meta, Resource},
-    client::APIClient,
-    config,
+    Client,
     runtime::Reflector,
 };
 
@@ -22,8 +21,7 @@ pub struct FooSpec {
 async fn main() -> anyhow::Result<()> {
     std::env::set_var("RUST_LOG", "info,kube=debug");
     env_logger::init();
-    let config = config::load_kube_config().await?;
-    let client = APIClient::new(config);
+    let client = Client::new().await?;
     let namespace = std::env::var("NAMESPACE").unwrap_or("default".into());
 
     // This example requires `kubectl apply -f examples/foo.yaml` run first